refactor(crypto): move normalizePubkey to curve_secp
